requirements.txt
colorama
python-dateutil
pyinstaller
